:sparkles: 优化脚下方块检测 :pencil: 优化下落测试代码 :bug: 修复1.12.2功能在1.16.2使用

+++ doctor-protocol/src/main/kotlin/top/fanua/doctor/protocol/definition/play/client/PlayerPositionAndLook.kt
@@ -66,9 +66,9 @@ data class PlayerPositionAndLookPacket(
 
 @Serializable
 data class Position(
-    val x: Long,
-    val y: Long,
-    val z: Long
+    val x: Int,
+    val y: Int,
+    val z: Int
 )
 
 class PlayerPositionAndLookDecoder : PacketDecoder<PlayerPositionAndLookPacket> {
